+++ R/ggforce_package.R
@@ -1,0 +1,3 @@
+#' @useDynLib ggforce
+#' @importFrom Rcpp sourceCpp
+NULL