+++ NEWS.md
@@ -5,3 +5,5 @@
 
 ## New features
 - geom_arc / stat_arc for drawing circle segments
+- geom_edge_bundle / stat_edge_bundle for drawing edge bundles based on control
+points